+++ src/main/java/com/back/BackApplication.java
@@ -1,13 +1,13 @@
-package com.example.demo;
+package com.back;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BackendApplication {
+public class BackApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(BackendApplication.class, args);
+        SpringApplication.run(BackApplication.class, args);
     }
 
 }
