--- src/main/java/com/example/demo/TestClass.java
@@ -1,9 +1,0 @@
-package com.example.demo;
-
-public class TestClass {
-
-    public int add(int a, int b) {
-        return a + b;
-    }
-
-}
