--- src/main/java/com/example/demo/order/controller/DummyController.java
@@ -1,4 +1,0 @@
-package com.example.demo.order.controller;
-
-public class DummyController {
-}
