--- src/main/java/com/example/demo/order/service/DummyService.java
@@ -1,4 +1,0 @@
-package com.example.demo.order.service;
-
-public class DummyService {
-}
