+++ src/main/java/com/example/order/controller/DummyController.java
@@ -1,0 +1,4 @@
+package com.example.order.controller;
+
+public class DummyController {
+}
