+++ src/main/java/com/example/order/service/DummyService.java
@@ -1,0 +1,4 @@
+package com.example.order.service;
+
+public class DummyService {
+}
